feat: add recent virus total scans to localhistory and restructure pages
