+++ src/src/components/Scanner/ScanURLForm.jsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from "react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { useRouter } from "next/navigation";
-import { Shield } from "lucide-react";
+import { saveToRecentScans } from "@/utils/scanHistory";
 
 export default function ScanUrlForm() {
   const [url, setUrl] = useState("");
@@ -38,7 +38,10 @@ export default function ScanUrlForm() {
         throw new Error("Failed to get analysis ID");
       }
 
-      // Instead of fetching results here, redirect to the results page
+      // Save to recent scans
+      saveToRecentScans("url", submitData.analysisId, url);
+
+      // Redirect to the results page
       router.push(`/scan/url/${submitData.analysisId}`);
     } catch (error) {
       console.error("Error:", error);
@@ -48,42 +51,29 @@ export default function ScanUrlForm() {
   };
 
   return (
-    <div className="mx-auto mt-10 flex w-full max-w-2xl flex-col items-center justify-center p-4">
-      <div className="text-primary mb-2">
-        <Shield size={48} />
-      </div>
-
-      <h1 className="mb-2 text-center text-3xl font-bold">URL Scanner</h1>
-
-      <p className="mb-6 text-center text-sm text-neutral-600">
-        Scan URLs, domains, and IP addresses to detect potential security
-        threats
+    <div className="w-full rounded-lg bg-white p-6 shadow-sm">
+      <h2 className="mb-1 text-xl font-medium">Enter a URL to scan</h2>
+      <p className="mb-4 text-sm text-neutral-600">
+        We'll check the URL against multiple security databases to identify
+        potential threats.
       </p>
 
-      <div className="w-full rounded-lg bg-white p-6 shadow-sm">
-        <h2 className="mb-1 text-xl font-medium">Enter a URL to scan</h2>
-        <p className="mb-4 text-sm text-neutral-600">
-          We'll check the URL against multiple security databases to identify
-          potential threats.
-        </p>
-
-        <div className="flex flex-col gap-4 sm:flex-row md:gap-2">
-          <Input
-            type="search"
-            placeholder="Enter URL To Scan"
-            value={url}
-            onChange={(e) => setUrl(e.target.value)}
-            disabled={loading}
-            className="flex-grow"
-          />
-          <Button
-            onClick={handleScan}
-            disabled={loading || !isValidUrl}
-            className="whitespace-nowrap"
-          >
-            {loading ? "Scanning..." : "Scan"}
-          </Button>
-        </div>
+      <div className="flex flex-col gap-4 sm:flex-row md:gap-2">
+        <Input
+          type="search"
+          placeholder="Enter URL To Scan"
+          value={url}
+          onChange={(e) => setUrl(e.target.value)}
+          disabled={loading}
+          className="flex-grow"
+        />
+        <Button
+          onClick={handleScan}
+          disabled={loading || !isValidUrl}
+          className="whitespace-nowrap"
+        >
+          {loading ? "Scanning..." : "Scan"}
+        </Button>
       </div>
     </div>
   );
